==> scripts/4-report.Rmd <==
---
title: "ANC4 & SAB coverage report"
author: "XXX"
date: "2025-07-27"
output:
  pdf_document:
  toc: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

# Clean the workspace
rm(list=ls()) 

# Download and install packages
list_packages <- c("tidyverse", "ARTofR", "openxlsx", "data.table", "here", "bookdown",
                  "ggplot2", "RColorBrewer", "dplyr", "knitr", "renv", "rmarkdown")

# If the package is not installed, then install it 
if (!require("pacman")) install.packages("pacman")
# Load the packages 
pacman::p_load(list_packages, character.only = TRUE, install = TRUE)

# Import the data
## here::here() is used so that paths are referred from project root
## Coverage of ANC4 and SAB for on-track and off-track categories
data_cover = read.csv(file = here::here("data/03_final/data_2022pop-weighted-coverage_anc4_sab_mortality-class.csv"))
## Coverage at country-level for ANC4 and SAB
## Used to assess data availability
data_combined = read.csv(file = here::here("data/02_clean/data_combined.csv")) %>% 
  # Create status "on-track" and "off-track" groups
  mutate(group = case_when(status == "Achieved" | status == "On Track" ~ "On-track",
                           status == "Acceleration Needed" ~ "Off-track"))
```

# Introduction

This short document presents the coverage of (1) the percentage of women (aged 15-49) with at least 4 antenatal care visits (ANC4); (2) the percentage of deliveries attended by skilled health personnel (SAB) for two categories of countries created from the under-five mortality classification. On-track countries are described as "achieved" or "on-track" while off-track countries are described as "off-track" under the classification.

This document is part of the assessment for the following positions at **UNICEF Data and Analytics** :

  * Learning and Skills Data Analyst Consultant (Req. #581598)
  * Household Survey Data Analyst Consultant (Req. #581656)
  * Administrative Data Analyst (Req. #581696)
  * Microdata Harmonization Consultant (Req. #581699)


# Results

The results of the analysis are presented on the figure below.

```{r, include = FALSE}
data_fig1 = data_cover %>%   
  # Rename for improved plotting
  rename("ANC4" = "anc4",
         "SAB" = "sab") %>% 
  # Remove useless variables
  #select(-c(n_countries, tot_pop_2022)) %>% 
  # Pivot longer for plotting
  pivot_longer(cols = c("ANC4", "SAB"),
               names_to = "indicator",
               values_to = "value") 

# Create the figure
fig1 = ggplot(data = data_fig1,
              aes(x = indicator, y = value, fill = group, label = round(value,1), group = group)) +
  geom_bar(position = "dodge", stat = "identity") +
  geom_text(aes(y = value+3),position = position_dodge(width=1)) +
  scale_fill_brewer(palette = "Blues") +
  labs(title = "Coverage of ANC4 and SBA across under-five mortality categories",
       subtitle = "On the period 2018-2022",
       x = "", y = "Population-weighted average",
       caption = paste0("Off-track (on-track) category covers ", data_fig1[data_fig1$group == "Off-track" & data_fig1$indicator == "ANC4",]$n_countries, " countries (", data_fig1[data_fig1$group == "On-track" & data_fig1$indicator == "ANC4",]$n_countries  ,") or ", format(data_fig1[data_fig1$group == "Off-track" & data_fig1$indicator == "ANC4",]$tot_pop_2022, big.mark = ",", digits = 0, scientific = F), " people (", format(data_fig1[data_fig1$group == "On-track" & data_fig1$indicator == "ANC4",]$tot_pop_2022, big.mark = ",", digits = 0, scientific = F),").\nANC4 : % of women (aged 15-49) with at least 4 antenatal care visits\nSAB : % of deliveries attended by skilled health personnel.\nFor both indicators we use the most recent available country estimate on the 2018-2022 period.\nIndicators are then aggrated using the 2022 projected population as weight. \nSources : UNICEF (population, ANC4 and SAB indicators, under-five mortality categories).")) +
  theme_minimal() +
  theme(plot.caption = element_text(hjust = 0),
        legend.title = element_blank())

```

```{r fig1, echo = FALSE, fig.align='center'}
# Show the figure
fig1

```

On the period 2018-2022, in off-track countries on average 46,8% of women had at least 4 antenatal care visits (ANC4) and 57,5% delivered with the attendance of skilled personnel (SAB). In on-track countries these figures are higher : 50,8% and 65,8%. This suggests better healthcare before birth and during delivery is correlated with lower under-five mortality rates. 

This analysis is very preliminary and an assessment of the factors impacting under-five mortality rates would require a deeper analysis (adding variables on mother living conditions before birth typically). Also this analysis is limited by availability of data on the 2018-2022 period : out of `r nrow(data_combined[data_combined$group == "On-track",])` on-track countries (`r nrow(data_combined[data_combined$group == "Off-track",])` off-track) `r nrow(data_combined[data_combined$group == "On-track" & is.na(data_combined$anc4) == F,])` (`r nrow(data_combined[data_combined$group == "Off-track" & is.na(data_combined$anc4) == F,])`) have an ANC4 estimate and `r nrow(data_combined[data_combined$group == "On-track" & is.na(data_combined$sab) == F,])` (`r nrow(data_combined[data_combined$group == "Off-track" & is.na(data_combined$sab) == F,])`) have a SAB estimate.
